Extract method for retrieving next opening day
